Override the font size scaling for smaller screens
On mobile, the font should be 150% of the default (18px default -> 24px on mobile). This makes reading the text easier.

I doubt that this is the correct way to override the `--pico-font-size` value only for screens less than 768px wide, but it seems to work.

I've copied the CSS code from pico.min.css for these values and put it in the style for the book page directly.